Question adding works now
But can only add a question then keep changing that question (need to
find a way to get unique keys that don’t break json)
