+++ app/build.gradle
@@ -44,6 +44,8 @@ dependencies {
     // volley http library
     implementation 'com.android.volley:volley:1.0.0'
     implementation 'com.google.code.gson:gson:2.6.2'
+    compile 'com.google.firebase:firebase-core:11.8.0'
+    compile 'com.google.firebase:firebase-database:11.8.0'
 }
 
 
@@ -55,3 +57,4 @@ allprojects {
         maven { url 'https://jitpack.io' }
     }
 }
+apply plugin: 'com.google.gms.google-services'
